\name{analogue-internal}
\alias{cumWmean}
\alias{cummean}
\alias{minDij}
\alias{maxBias}
\alias{.simpleCap}
\alias{wmean}
\alias{w.avg}
\alias{inv.deshrink}
\alias{class.deshrink}
\alias{expand.deshrink}
\alias{no.deshrink}
\alias{RowSums}
\alias{ColSums}
\alias{deshrink.pred}
\title{Internal analogue Functions}
\description{
  Internal analogue functions
}
\details{
  These are not to be called by the user.
}
\keyword{ internal }
