\name{analogue-package}
\alias{analogue-package}
\alias{analogue}
\docType{package}
\title{
Analogue methods for palaeoecology
}
\description{
Fits Modern Analogue Technique transfer function models
for prediction of environmental data from species data. Also
performs analogue matching, a related technique used in palaeo
ecological restoration.
}
\details{
\tabular{ll}{
Package: \tab analogue\cr
Type: \tab Package\cr
Version: \tab 0.2-7\cr
Date: \tab 2006-12-06\cr
Depends: \tab \cr
License: \tab GPL Version 2\cr
Encoding: \tab UTF-8\cr
Packaged: \tab Wed Dec 6 16:33:51 2006; gavin\cr
Built: \tab R 2.4.0; ; 2006-12-06 16:33:56; unix\cr
}

Index:
\preformatted{
Screeplot               Screeplots of model results
analog                  Analogue matching
bayesF                  Bayes factors
bootstrap               Bootstrap estimation and errors
cma                     Close modern analogues
dissimilarities         Extract dissimilarity coefficients from models
distance                Flexibly calculate dissimilarity or distance
                        measures
fitted.mat              Fitted values of a MAT model
join                    Merge species data sets on common columns
                        (species)
mat                     Modern Analogue Technique transfer function
                        models
mcarlo                  Monte Carlo simulation of dissimilarities
plot.bayesF             Bayes factor plots
plot.cma                Draws stripcharts of close modern analogue
                        results
plot.dissimilarities    Plots the distribution of extracted
                        dissimilarities
plot.mat                Plot diagnostics for a mat object
plot.mcarlo             Plotting of Monte Carlo simulated dissimilarity
                        distributions
plot.roc                Plot ROC curves and associated diagnostics
predict.mat             Predict method for Modern Analogue Technique
                        models
reconPlot               Stratigraphic plots of palaeoenvironmental
                        reconstructions
residuals.bootstrap.mat
                        Extract Bootstrapped Model Residuals
residuals.mat           Extract Model Residuals
rlgh                    Round Loch of Glenhead Diatoms
roc                     ROC curve analysis
summary.analog          Summarising analogue matching results
summary.bootstrap.mat   Summarising bootstrap resampling for MAT models
summary.cma             Summarising the extraction of close modern
                        analogues
summary.mat             Summarising Modern Analogue Technique models
summary.predict.mat     Summarising MAT model predictions
swapdiat                SWAP sub-fossil diatom and pH training set
swappH                  SWAP sub-fossil diatom and pH training set
}
}
\author{
Gavin L. Simpson.

Maintainer: Gavin L. Simpson <gavin.simpson@ucl.ac.uk>
}
\keyword{ package }
