\name{rlgh}
\alias{rlgh}
\docType{data}
\title{Round Loch of Glenhead Diatoms}
\description{
  Diatom sub-fossil assemblage counts from the Round Loch of Glenhead.
}
\usage{data(rlgh)}
\format{
  A data frame with 101 observations on 139 diatom species.
}
\details{
  The samples are taken from various depths down a sediment core
  retrieved from the Round Loch of Glenhead, Galloway, Scotland.
}
\references{

  Jones, V.J., Stevenson, A.C. and Battarbee, R.W. (1989) Acidification
  of lakes in Galloway, south west Scotland --- A diatom and pollen
  study of the post-glacial history of the Round Loch of
  Glenhead. \emph{Journal of Ecology} \strong{77(1)}, 1--23.
}
\examples{
data(rlgh)
}
\keyword{datasets}
