\name{swappH}
\alias{swappH}
\docType{data}
\title{SWAP sub-fossil diatom and pH training set}
\description{
  The Surface Waters Acidifcation Project (SWAP) Palaeolimnology
  Programme diatom-pH surface sample training set.
}
\usage{data(swappH)}
\format{
  Numeric vector containing the pH of 74 lakes from the SWAP diatom-pH
  training set.
}
\references{

  Stevenson, A.C., Juggins, S., Birks, H.J.B., Anderson, D.S., Anderson,
  N.J., Battarbee, R.W., Berge, F., Davis, R.B., Flower, R.J., Haworth,
  E.Y., Jones, V.J., Kingston, J.C., Kreiser, A.M., Line, J.M., Munro,
  M.A.R., and Renberg, I. (1995). \emph{The Surface Waters Acidification
  Project Palaeolimnology programme: modern diatom/lake-water chemistry
  data-set}. ENSIS Publishing, London.

}
\seealso{\code{\link{swapdiat}}}
\examples{
data(swappH)
str(swappH)
}
\keyword{datasets}
