\name{analogue-package}
\alias{analogue-package}
\alias{analogue}
\docType{package}
\title{
Analogue methods for palaeoecology
}
\description{
Fits Modern Analogue Technique transfer function models
for prediction of environmental data from species data. Also
performs analogue matching, a related technique used in palaeo
ecological restoration.
}
\details{
\tabular{ll}{
Package: \tab analogue\cr
Type: \tab Package\cr
Version: \tab 0.4-3\cr
Date: \tab 2007-09-15\cr
Depends: \tab R (>= 2.5.0), stats, graphics, vegan\cr
License: \tab GPL Version 2\cr
Built: \tab R 2.5.1; ; 2007-08-03 13:44:47; unix\cr
}

Index:
\preformatted{
RMSEP                   Root mean square error of prediction
analog                  Analogue matching
analogue-package        Analogue methods for palaeoecology
bayesF                  Bayes factors
bootstrap               Bootstrap estimation and errors
bootstrapObject         Bootstrap object description
cma                     Close modern analogues
dissimilarities         Extract dissimilarity coefficients from models
distance                Flexibly calculate dissimilarity or distance
                        measures
getK                    Extract and set the number of analogues
join                    Merge species data sets on common columns
                        (species)
mat                     Modern Analogue Technique transfer function
                        models
mcarlo                  Monte Carlo simulation of dissimilarities
minDC                   Extract minimum dissimilarities
plot.bayesF             Bayes factor plots
plot.dissimilarities    Plots the distribution of extracted
                        dissimilarities
plot.mat                Plot diagnostics for a mat object
plot.mcarlo             Plott Monte Carlo simulated dissimilarity
                        distributions
plot.minDC              Plot of minimum dissimilarity per sample
plot.roc                Plot ROC curves and associated diagnostics
predict.mat             Predict method for Modern Analogue Technique
                        models
reconPlot               Stratigraphic plots of palaeoenvironmental
                        reconstructions
rlgh                    Round Loch of Glenhead Diatoms
roc                     ROC curve analysis
screeplot.mat           Screeplots of model results
summary.analog          Summarise analogue matching results
summary.bootstrap.mat   Summarise bootstrap resampling for MAT models
summary.cma             Summarise the extraction of close modern
                        analogues
summary.mat             Summarise Modern Analogue Technique models
summary.predict.mat     Summarise MAT model predictions
swapdiat                SWAP sub-fossil diatom training set
swappH                  SWAP pH training set
}
}
\author{
Gavin L. Simpson.

Maintainer: Gavin L. Simpson <gavin.simpson@ucl.ac.uk>
}
\keyword{ package }
